Remove need for weird output utils
