--- lib/lev/utils/output_sources.rb
@@ -1,2 +1,0 @@
-require 'lev/utils/output_sources/attribute_subroutines'
-require 'lev/utils/output_sources/verbatim_subroutines'
