--- lib/lev/utils/output_sources/verbatim_subroutines.rb
@@ -1,30 +1,0 @@
-module Lev
-  module Utils
-    module OutputSources
-      class VerbatimSubroutines
-        def self.setup(routine_class, map)
-          map.values.each do |source|
-            routine_class.subroutines.add(source)
-            promote_verbatim_attributes(routine_class, nil, source)
-          end
-        end
-
-        private
-        def self.promote_verbatim_attributes(routine_class, key, source)
-          [source].flatten.each do |src|
-            key ||= Symbolify.exec(src)
-
-            nested_class = Nameify.exec(src)
-            source_class = routine_class.subroutines.routine_class(key)
-
-            nested_class.explicit_outputs.each do |attr|
-              routine_class.subroutines.add_attribute(key, attr)
-            end
-
-            promote_verbatim_attributes(routine_class, key, nested_class.verbatim_outputs)
-          end
-        end
-      end
-    end
-  end
-end
